test(obligation): add reset function for obligation import statistics
Implemented a reset function to clear obligation import statistics before each test case, ensuring accurate tracking of import metrics during integration tests.
