+++ tests/importer/talent-die-modifiers.spec.mjs
@@ -3,11 +3,10 @@
  * Couvre l'extraction, la normalisation et le formatage des modificateurs de dés
  */
 
-import { describe, it, expect, beforeEach } from 'vitest'
+import { describe, it, expect } from 'vitest'
 import {
   extractTalentDieModifiers,
   formatTalentDieModifiersForDescription,
-  sanitizeTalentDescription,
   extractTalentSource,
   assembleTalentDescription,
 } from '../../module/importer/mappings/oggdude-talent-diemodifiers-map.mjs'
@@ -246,58 +245,6 @@ describe('formatTalentDieModifiersForDescription', () => {
     const result = formatTalentDieModifiersForDescription(modifiers)
 
     expect(result).toContain('Upgrade ability 1 time')
-  })
-})
-
-describe('sanitizeTalentDescription', () => {
-  it('devrait nettoyer les balises script', () => {
-    const description = 'Normal text <script>alert("XSS")</script> more text'
-
-    const result = sanitizeTalentDescription(description)
-
-    expect(result).toBe('Normal text more text')
-    expect(result).not.toContain('script')
-    expect(result).not.toContain('alert')
-  })
-
-  it('devrait nettoyer les balises style', () => {
-    const description = 'Normal text <style>body{display:none}</style> more text'
-
-    const result = sanitizeTalentDescription(description)
-
-    expect(result).toBe('Normal text more text')
-    expect(result).not.toContain('style')
-  })
-
-  it('devrait normaliser les espaces multiples', () => {
-    const description = 'Text    with     multiple     spaces'
-
-    const result = sanitizeTalentDescription(description)
-
-    expect(result).toBe('Text with multiple spaces')
-  })
-
-  it('devrait tronquer à la longueur maximale', () => {
-    const description = 'A'.repeat(2500)
-
-    const result = sanitizeTalentDescription(description, 2000)
-
-    expect(result.length).toBe(2000)
-    expect(result).toMatch(/\.\.\.$/u)
-  })
-
-  it('devrait retourner une chaîne vide pour entrée nulle', () => {
-    expect(sanitizeTalentDescription(null)).toBe('')
-    expect(sanitizeTalentDescription(undefined)).toBe('')
-    expect(sanitizeTalentDescription('')).toBe('')
-  })
-
-  it('devrait trim les espaces au début et à la fin', () => {
-    const description = '   Text with spaces   '
-
-    const result = sanitizeTalentDescription(description)
-
-    expect(result).toBe('Text with spaces')
   })
 })
